Started work on an exec-based version Pangolin
Each of the Jar, Javac and Junit classes have been moved to a directory called "java", and each will get a double in a directory called "exec". The appropriate version will be require'd depending on the platform. The common functionality will be in a mixin that resides in a directory called "common".

So far the Jar feature has been ported -- it required minimal changes, although it required a dependency on the rubyzip gem.